% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentAbundances.R
\name{abundance}
\alias{abundance}
\title{Rodent Abundance}
\usage{
abundance(path = "~", level = "Site", type = "Rodents", length = "all",
  unknowns = F, incomplete = F, shape = "crosstab", time = "period")
}
\arguments{
\item{path}{path to location of downloaded Portal data; or 'repo' to retrieve data from github repo}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species; either all "Rodents" or only "Granivores"}

\item{length}{specify subset of plots; use "All" plots or only "Longterm" plots (plots that have had same treatment for entire time series)}

\item{unknowns}{either removes all individuals not identified to species (unknowns=F) or sums them in an additional column (unknowns=T)}

\item{incomplete}{either removes all data from incomplete trapping sessions (incomplete = F) or includes them (incomplete = T)
[note that if level="plot" and incomplete=T, NAs will be included in periods where trapping was incomplete]}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{time}{return data using the complete "newmoon" numbers of the original "period" numbers}
}
\description{
A function to summarize monthly Portal rodent species abundance
}
