% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_newmoon_code}
\alias{add_newmoon_code}
\title{Add NewMoon Codes}
\usage{
add_newmoon_code(summary_table, newmoon_table, time)
}
\arguments{
\item{summary_table}{Data.table with summarized rodent data.}

\item{newmoon_table}{Data_table linking newmoon codes with period codes.}

\item{time}{Character. Denotes whether newmoon codes are desired.}
}
\value{
Data.table of summarized rodent data with period or newmoon code
}
\description{
period codes denote the number of censuses that have occurred, but are
not the same as the number of censuses that should have occurred. Sometimes
censuses are missed (weather, transport issues,etc). You can't pick this
up with the period code. Because censues may not always occur monthly due to
the newmoon -  a new moon code was devised to give a standardized language
of time for forcasting in particular.
}
