% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_time}
\alias{add_time}
\title{Add User-specified time column}
\usage{
add_time(summary_table, newmoon_table, time = "period")
}
\arguments{
\item{summary_table}{Data.table with summarized rodent data.}

\item{newmoon_table}{Data_table linking newmoon codes with period codes.}

\item{time}{Character. Denotes whether newmoon codes, period codes,
and/or date are desired.}
}
\value{
Data.table of summarized rodent data with user-specified time format
}
\description{
period codes denote the number of censuses that have occurred, but are
not the same as the number of censuses that should have occurred. Sometimes
censuses are missed (weather, transport issues,etc). You can't pick this
up with the period code. Because censues may not always occur monthly due to
the newmoon -  a new moon code was devised to give a standardized language
of time for forcasting in particular. This function allows the user to decide
if they want to use the rodent period code, the new moon code, the date of 
the rodent census, or have their data with all three time formats
}
