% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggTime}
\alias{aggTime}
\title{Aggregate PORTAL data over time.}
\usage{
aggTime(unAgg, aggVars, timeCut, hod = TRUE)
}
\arguments{
\item{unAgg}{Data that you want to aggregate.}

\item{aggVars}{Variables you want to retain in aggregation, such as 'lanenumber' or 'stationid'}

\item{timeCut}{A string indicating how you want to aggregate over time. Examples include '1 min', '5 min', '1 hour', '1 day', etc.}

\item{hod}{A boolean indicating weather you want to aggregate by hour of day across multiple days.}
}
\value{
Aggregated data.
}
\description{
Aggregate PORTAL data over time.
}

