% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant_data_summaries.R
\name{bait_presence_absence}
\alias{bait_presence_absence}
\title{Ant Bait Presence Absence}
\usage{
bait_presence_absence(level = "Site")
}
\arguments{
\item{level}{level at which to summarize data: 'Site' or 'Plot'}
}
\value{
data frame with year, species, (plot if applicable), and presence [1,0]
}
\description{
Get ant species presence/absence by year (and plot if desired) from bait census data

Bait census data is more consistent over time than the colony census data. This function assumes that all species
present in at least one census were censused in all years.
}
