% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate census gaps}
\usage{
calc_diff(data)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}
}
\value{

}
\description{
Calculates the number of months between subsequent
censuses (diff) and adds to data.frame
}
