% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{check_for_newer_data}
\alias{check_for_newer_data}
\title{Check for latest version of data files}
\usage{
check_for_newer_data(newmoons_table, release_only = TRUE)
}
\arguments{
\item{newmoons_table}{the read in data.frame from
"PortalData/Rodents/moon_dates.csv"}

\item{release_only}{whether to check against the latest release or against
the latest commit on GitHub (defaults to TRUE)}
}
\value{
bool TRUE if there is a newer version of the data online
}
\description{
Check the latest newmoonnumber against the data that exists on
  the web (as of now, this means against the latest release or latest commit
  on GitHub)
}
