% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clean_data}
\alias{clean_data}
\title{Remove data still under QA/QC.}
\usage{
clean_data(full_data, trapping_table)
}
\arguments{
\item{full_data}{Data.table of raw data.}

\item{trapping_table}{Data.table with QCflag}
}
\value{
Data.table with latest 12 months of data removed.
}
\description{
Removes records that are still undergoing QA/QC (latest 12 months)
}
