% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clean_plant_data}
\alias{clean_plant_data}
\title{Do basic cleaning of Portal plant data}
\usage{
clean_plant_data(data_tables, type = "All", unknowns = FALSE, correct_sp,
  length = "all")
}
\arguments{
\item{data_tables}{the list of data_tables, returned from calling
\code{\link{loadPlantData}}}

\item{type}{specify subset of species; either "All" - includes annuals,
perennials, and shrubs; "Annuals" - only annuals; or "Non-woody" - includes
annuals and perennials but not shrubs}

\item{unknowns}{either removes all individuals not identified to species
(unknowns = FALSE) or sums them in an additional column (unknowns = TRUE)}

<<<<<<< 6314ff373f0dd9ebe173c0025e13b9e6010ac7ad
\item{correct_sp}{T/F whether or not to use likely corrected plant IDs, passed to \code{rename_species_plants}}

=======
>>>>>>> Update docs and tests
\item{length}{specify subset of plots; use "All" plots or only "Longterm"
plots (plots that have had same treatment for entire time series)}
}
\description{
This function does basic quality control of the Portal plant
  data. It is mainly called from \code{\link{get_plant_data}}, with
  several arguments passed along.

  The specific steps it does are, in order:
    (1) remove records with "bad" period codes or plot numbers via
        \code{\link{remove_suspect_entries}}
    (2) restrict species to annuals or non-woody via
        \code{link{process_annuals}}
    (3) remove records for unidentified species via
        \code{\link{process_unknownsp_plants}}
    (4) exclude incomplete trapping sessions via
        \code{\link{remove_incomplete_censuses}}
    (5) exclude the plots that aren't long-term treatments via
        \code{\link{filter_plots}}
}
