% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant_data_summaries.R
\name{colony_presence_absence}
\alias{colony_presence_absence}
\title{Ant Colony Presence Absence}
\usage{
colony_presence_absence(level = "Site", rare_sp = F)
}
\arguments{
\item{level}{level at which to summarize data: 'Site', 'Plot', or 'Stake'}

\item{rare_sp}{include rare species (T) or not (F).
Rare species may or may not have been censused in all years. Setting rare_sp=F gives a more conservative estimate of presence/absence}
}
\value{
data frame with year, species, (plot if applicable), and presence [1,0,NA]
}
\description{
Get ant species presence/absence by year (and plot if desired) from colony census data

Anomalies in ant colony census protocol over the years means that it can be difficult to discern true absences
of all species in all years.  This function uses information from Portal_ant_species.csv and Portal_ant_dataflags.csv
to predict true presence/absence of species per plot per year.  If a more conservative estimate is desired,
setting the argument 'rare_sp = T' will only include species we are confident were censused regularly. Setting
'rare_sp = F' may include some false absences, since it is unknown if some rare species were censused in all years.
Unknowns are always xcluded, as well as extremely rare species (if a species is only recorded a few times in one year of 30,
it was assumed that data for this species was not collected consistently through time)
}
