% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{download_observations}
\alias{download_observations}
\title{Download the PortalData repo}
\usage{
download_observations(base_folder = "~")
}
\value{
None
}
\description{
This downloads the latest portal data regardless if they are
actually updated or not.
}
\details{
TODO: incorporate data retriever into this when it's pointed at the github repo
}
