% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI.R
\name{fill_missing_ndvi}
\alias{fill_missing_ndvi}
\title{Fill in historic ndvi data to the complete timeseries being fit}
\usage{
fill_missing_ndvi(ndvi, level, last_time, moons = NULL)
}
\arguments{
\item{ndvi}{ndvi data}

\item{level}{specify "monthly" or "newmoon"}

\item{last_time}{the last time step to have been completed}

\item{moons}{moon data (required if level = "newmoons" and forecasts are
needed)}
}
\value{
a data.frame with time and ndvi values
}
\description{
Fill in historic ndvi data to the complete timeseries being fit
}
\details{
missing values during the time series are replaced using
 na.interp, missing values at the end of the time series are forecast using
  auto.arima with seasonality (using Fourier transform)
}
