% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{filter_diffs}
\alias{filter_diffs}
\title{Filter period codes to match gaps between censuses}
\usage{
filter_diffs(data, diff_value)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}

\item{diff_value}{the number of months between subsequent censuses}
}
\value{

}
\description{
Returns the period code for months that match the specified diff_value.
Diffs are the number of months between subsequent censuses
}
