% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filtering.R
\name{filterFreeway}
\alias{filterFreeway}
\title{Filter freeway data using threshold values}
\usage{
filterFreeway(fdata)
}
\arguments{
\item{fdata}{Your raw freeway data.}
}
\value{
Filtered freeway data
}
\description{
This function will use six criteria developed for filtering PORTAL data to filter your raw data. The criteria are summmarized on Table 1 of this document: http://portal.its.pdx.edu/Portal/static/files/fhwa/Freeway%20Data%20Documentation.pdf
}

