% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{fix_period}
\alias{fix_period}
\title{Fix Period Date}
\usage{
fix_period(data, i_period, adj)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}

\item{i_period}{period to fix}

\item{adj}{Months to adjust date by}
}
\value{

}
\description{
Returns modified monthly date for a period.
}
