% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format-message.R
\name{format_code}
\alias{format_code}
\title{Format content as code}
\usage{
format_code(...)
}
\arguments{
\item{...}{strings}
}
\value{
a formatted string to output
}
\description{
Format content as code
}
