% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format-message.R
\name{format_todo}
\alias{format_todo}
\title{Format content as an action to be performed by the user}
\usage{
format_todo(...)
}
\arguments{
\item{...}{strings}
}
\value{
a formatted string to output
}
\description{
Format content as an action to be performed by the user
}
