% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format-message.R
\name{format_value}
\alias{format_value}
\title{Format content as a variable value}
\usage{
format_value(...)
}
\arguments{
\item{...}{strings}
}
\value{
a formatted string to output
}
\description{
Format content as a variable value
}
