% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/querying.R
\name{freewayData}
\alias{freewayData}
\title{Query and return freeway loop and radar data from PORTAL}
\usage{
freewayData(con, dets, startTime, endTime)
}
\arguments{
\item{dets}{Vector of detector IDs.}

\item{startTime}{Start time of requested data in either a YYYY-MM-DD or YYYY-MM-DD HH:MM:SS format.}

\item{endTime}{End time of requested data in either a YYYY-MM-DD or YYYY-MM-DD HH:MM:SS format.}
}
\value{
Result of freeway data query.
}
\description{
Query and return freeway loop and radar data from PORTAL
}

