% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{freewaySpeedMap}
\alias{freewaySpeedMap}
\title{freewaySpeedMap}
\usage{
freewaySpeedMap(con, corridorID, startDate, endDate, weekdays = TRUE,
  outputPng = NULL)
}
\arguments{
\item{con}{database connection PORTAL PostgreSQL database}

\item{corridorID}{ID number of the corridor to plot. See 'corrdidors' table to select ID for plotting.}

\item{startDate}{Start date of data to query (YYYY-MM-DD format)}

\item{endDate}{End date of data to query (YYYY-MM-DD format)}

\item{weekdays}{Boolean indicating whether to subset data to weekdays. Defaults to TRUE.}

\item{outputPng}{.png file path of output plot. Default is NULL; if NULL plots to current device (i.e. RStudio plot device).}
}
\value{
None
}
\description{
Produces a speed-based heatmap (AKA brainscan plot) for a corridor with PORTAL's freeway data system. See Freeway Speed Map vignette for example usage.
}

