% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{full_path}
\alias{full_path}
\title{Full Path}
\usage{
full_path(ReferencePath, BasePath = getwd())
}
\arguments{
\item{ReferencePath}{a path to join with current working directory}

\item{BasePath}{Current working directory else path given}
}
\value{
Full path
}
\description{
Return normalized path for all operating systems
}
\examples{
full_path('PortalData/Rodents/Portal_rodent.csv')
full_path('PortalData/Rodents/Portal_rodent.csv', '~')
}
