% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{get_data_versions}
\alias{get_data_versions}
\title{get version and download info for PortalData}
\usage{
get_data_versions(from_zenodo = FALSE, halt_on_error = FALSE)
}
\arguments{
\item{from_zenodo}{logical; if `TRUE`, get info from Zenodo, otherwise GitHub}

\item{halt_on_error}{logical; if `FALSE`, return NULL on errors, otherwise
whatever got returned (could be an error or warning)}
}
\value{
A data.frame with two columns, `version` (string with the version #) and
  `zipball_url` (download URLs for the corresponding zipped release).
}
\description{
Check either Zenodo or GitHub for the version and download link
  for PortalData.
}
