% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portalr-package.R
\name{get_dataset_citation}
\alias{get_dataset_citation}
\title{Return Citation for Portal Data}
\usage{
get_dataset_citation()
}
\value{
An object of class "citation". For more details, see `citation()`
}
\description{
Return Citation for Portal Data
}
