% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_moons.R
\name{get_future_moons}
\alias{get_future_moons}
\title{Get next 12 new moon dates and assign newmoon numbers for forecasting}
\usage{
get_future_moons(moons, num_future_moons = 12)
}
\arguments{
\item{moons}{current newmoonnumber table}

\item{num_future_moons}{number of future moons to get}
}
\value{
expected moons table for 12 future new moons
}
\description{
Get next 12 new moon dates and assign newmoon numbers for forecasting
}
