% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_moons.R
\name{get_future_newmoons}
\alias{get_future_newmoons}
\title{Get future newmoon dates and numbers}
\usage{
get_future_newmoons(newmoons, nfuture_newmoons = NULL)
}
\arguments{
\item{newmoons}{current newmoon table}

\item{nfuture_newmoons}{number of future newmoons to get}
}
\value{
expected newmoons table for requested future newmoons
}
\description{
Get next newmoon dates and assign newmoon numbers for forecasting
}
