% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{get_github_releases}
\alias{get_github_releases}
\title{get GitHub Release Info for PortalData}
\usage{
get_github_releases()
}
\value{
A data.frame with two columns, `tag_name` (name of the tags) and
  `zipball_url` (download URLs for the corresponding zipped release)
}
\description{
Use the GitHub API to get info about the releases of the
  PortalData repo.
}
