% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantData.R
\name{get_plant_data}
\alias{get_plant_data}
\alias{plant_abundance}
\title{Generate summaries of Portal plant data}
\usage{
get_plant_data(path = "~", level = "Site", type = "All", length = "all",
  unknowns = FALSE, correct_sp = TRUE, shape = "crosstab",
  output = "abundance")

plant_abundance(...)
}
\arguments{
\item{path}{path to location of downloaded Portal data; or "repo" to
retrieve data from github repo}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species; either "All" - includes annuals,
perennials, and shrubs; "Annuals" - only annuals; or "Non-woody" - includes
annuals and perennials, not shrubs or subshrubs}

\item{length}{specify subset of plots; use "All" plots or only "Longterm"
plots (plots that have had same treatment for entire time series)}

\item{unknowns}{either removes all individuals not identified to species
(unknowns = FALSE) or sums them in an additional column (unknowns = TRUE)}

\item{correct_sp}{correct species names suspected to be incorrect in early data (T/F)}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{output}{specify whether to return "abundance", or "cover" [cover data
starts in summer 2015]}

\item{...}{arguments passed to \code{\link{get_plant_data}}}
}
\value{
a data.frame in either "long" or "wide" format, depending on the
  value of `shape`
}
\description{
This function is a generic interface into creating
  summaries of the Portal plant species data. It contains a number of
  arguments to specify both the kind of data to summarize, at what level of
  aggregation, various choices for dealing with data quality, and output
  format.

\code{plant_abundance} generates a table of plant abundance
}
\examples{
plant_abundance("repo")

}
