% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{get_rodent_data}
\alias{get_rodent_data}
\alias{abundance}
\alias{biomass}
\alias{energy}
\alias{abundance}
\alias{biomass}
\alias{energy}
\title{Generate summaries of Portal rodent data}
\usage{
get_rodent_data(path = "~", level = "Site", type = "Rodents",
  length = "all", unknowns = FALSE, incomplete = FALSE,
  shape = "crosstab", time = "period", output = "abundance",
  fillweight = (output != "abundance"))

abundance(...)

biomass(...)

energy(...)
}
\arguments{
\item{path}{path to location of downloaded Portal data; or "repo" to
retrieve data from github repo}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species; either all "Rodents" or only
"Granivores"}

\item{length}{specify subset of plots; use "All" plots or only "Longterm"
plots (plots that have had same treatment for entire time series)}

\item{unknowns}{either removes all individuals not identified to species
(unknowns = FALSE) or sums them in an additional column (unknowns = TRUE)}

\item{incomplete}{either removes all data from incomplete trapping sessions
(incomplete = FALSE) or includes them (incomplete = TRUE)
[note that if level="plot" and incomplete=T, NAs will be included in periods where trapping was incomplete]}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{time}{specify the format of the time index in the output, either
"period" (sequential Portal surveys), "newmoon" (lunar cycle numbering),
"date" (calendar date)}

\item{output}{specify whether to return "abundance", or "biomass", or "energy"}

\item{fillweight}{specify whether to fill in unknown weights with other records from that individual or species, where possible}

\item{...}{arguments passed to \code{\link{get_rodent_data}}}
}
\description{
This function is a generic interface into creating (monthly)
  summaries of the Portal rodent species data. It contains a number of
  arguments to specify both the kind of data to summarize, at what level of
  aggregation, various choices for dealing with data quality, and output
  format.

\code{abundance} generates a table of rodent abundance

* \code{biomass()} generates a table of rodent biomass

* \code{energy()} generates a table of rodent energy (computed as biomass ^ 0.75)
}
\examples{
abundance("repo")

biomass("repo")

energy("repo")

}
