% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{gh_token}
\alias{gh_token}
\title{Return the local user's GitHub Personal Access Token (PAT) (from gh package)
https://github.com/r-lib/gh/blob/master/R/gh_request.R}
\usage{
gh_token()
}
\value{
A string, with the token, or a zero length string scalar,
if no token is available.
}
\description{
Return the local user's GitHub Personal Access Token (PAT) (from gh package)
https://github.com/r-lib/gh/blob/master/R/gh_request.R
}
