% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{insert_missing_months}
\alias{insert_missing_months}
\title{Insert Missing Months}
\usage{
insert_missing_months(data, i_period)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}

\item{i_period}{period to fix}
}
\value{

}
\description{
Insert missing month
}
