% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_census_to_dates}
\alias{join_census_to_dates}
\title{Join census, dates, and plot treatment tables}
\usage{
join_census_to_dates(census_table, date_table, plots_table)
}
\arguments{
\item{census_table}{Data_table of plant censuses}

\item{date_table}{Data table of dates of plant censuses}

\item{plots_table}{Data_table of treatments for the plots.}
}
\value{
Data.table of quadrat data with treatment info added.
}
\description{
Joins plant census table, census date table, and plot treatment tables
}
