% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_census_to_quadrats}
\alias{join_census_to_quadrats}
\title{Join quadrat and census tables}
\usage{
join_census_to_quadrats(quadrat_data, census_table)
}
\arguments{
\item{quadrat_data}{Data.table with raw rodent data.}

\item{census_table}{Data_table of when plots were censused.}
}
\value{
Data.table of raw quadrat data with census info added.
}
\description{
Joins quadrat data with list of census dates
}
