% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_plots_to_rodents}
\alias{join_plots_to_rodents}
\title{Join rodent and plot tables}
\usage{
join_plots_to_rodents(rodent_table, plots_table)
}
\arguments{
\item{plots_table}{Data_table of treatments for the plots.}

\item{rodent_data}{Data.table with raw rodent data.}
}
\value{
Data.table of raw rodent data with treatment info added.
}
