% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{join_plots_to_trapping}
\alias{join_plots_to_trapping}
\title{Join plots and trapping tables}
\usage{
join_plots_to_trapping(trapping, plots)
}
\arguments{
\item{trapping}{trapping_table from Portal data (can be filtered)}

\item{plots}{plots_table from Portal data}
}
\value{
trapping table with sampled column removed and treatment column 
  added
}
\description{
Join plots and trapping tables
}
