% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_trapping_to_rodents}
\alias{join_trapping_to_rodents}
\title{Join rodent and trapping tables}
\usage{
join_trapping_to_rodents(rodent_table, trapping_table, incomplete)
}
\arguments{
\item{trapping_table}{Data_table of when plots were censused.}

\item{rodent_data}{Data.table with raw rodent data.}
}
\value{
Data.table of raw rodent data with trapping info added.
}
