% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{load_ant_data}
\alias{load_ant_data}
\title{Loads Portal ant data files.}
\usage{
load_ant_data(path = "~", download_if_missing = TRUE)
}
\arguments{
\item{path}{either the file path that contains the PortalData folder or
"repo", which then pulls data from the PortalData GitHub repository}

\item{download_if_missing}{if the specified file path doesn't have the
PortalData folder, then download it}
}
\value{
List of 4 dataframes:
  \itemize{
    \item bait_data. raw ant bait data
    \item colony_data. raw ant colony data
    \item species_table. species code, names, types
    \item plots_table. treatment assignments for each plot.
  }
}
\description{
Loads main Portal ant data files from either
a user defined path or the online Github repository.
}
\examples{
portal_ant_data <- load_ant_data("repo")

}
