% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{load_plant_data}
\alias{load_plant_data}
\title{Loads Portal plant data files.}
\usage{
load_plant_data(path = "~", download_if_missing = TRUE)
}
\arguments{
\item{path}{either the file path that contains the PortalData folder or
"repo", which then pulls data from the PortalData GitHub repository}

\item{download_if_missing}{if the specified file path doesn't have the
PortalData folder, then download it}
}
\value{
List of 5 dataframes:
  \itemize{
    \item quadrat_data. raw plant quadrat data
    \item species_table. species code, names, types
    \item census_table. indicates whether each quadrat was counted in each
      census; area of each quadrat
    \item date_table. start and end date of each plant census
    \item plots_table. rodent treatment assignments for each plot.
  }
}
\description{
Loads main Portal plant data files from either
a user defined path or the online Github repository.
}
\examples{
portal_plant_data <- load_plant_data("repo")

}
