% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_plant_pata}
\alias{load_plant_pata}
\title{Loads Portal plant data files.}
\usage{
load_plant_pata(path = "~")
}
\arguments{
\item{path}{string Containing path to PortalData folder
should include ending /; if path = 'repo',
data is pulled from PortalData GitHub repository.}
}
\value{
List of 4 dataframes:
\itemize{
\item quadrat_data. raw plant quadrat data
\item species_table. species code, names, types
\item census_table. indicates whether each quadrat was counted in each census; area of each quadrat
\item date_table. start and end date of each plant census
\item plots_table. rodent treatment assignments for each plot.
}
}
\description{
Loads main Portal plant data files from either
a user defined path or the online Github repository.
}
\examples{
portal_plant_data <- load_plant_data("repo")
}
