% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{loadData}
\alias{loadData}
\title{Loads Portal rodent data files.}
\usage{
loadData(path)
}
\arguments{
\item{path}{string Containing path to PortalData folder
should include ending /; if path = 'repo', 
data is pulled from PortalData GitHub repository.}
}
\value{
List of 5 dataframes:
\itemize{
\item rodent_data. raw data on rodent captures
\item species_table. species code, names, types
\item trapping_table. when each plot was trapped
\item newmoons_table. pairs census periods with newmoons
\item plots_table. rodent treatment assignments for each plot.
}
}
\description{
Loads main Portal rodent data files from either
a user defined path or the online Github repository.
}
\examples{
loadData('repo')
loadData('../')

}
