% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{make_level_data}
\alias{make_level_data}
\title{Rodent data summarized at the relevant level (plot, treatment, site)}
\usage{
make_level_data(plot_data, level, output, min_plots)
}
\arguments{
\item{plot_data}{rodent data summarized at the plot level}

\item{level}{specify level of interest ("plot", "treatment", "site")}

\item{output}{specify whether to return "abundance", or "biomass",
or "energy"}

\item{min_plots}{minimum number of plots within a period for an
observation to be included}
}
\value{
fully crossed flat table of observations with effort (number of
  traps and number of plots). The crossing depends on the level:
  "plot" is period x treatment x plot x species, "treatment" is
  period x treatment x species, and "site" is period x species. Any
  observations not sufficiently (as defined by min_plots, and
  hierarchically by min_traps) sampled are returned with NA
  for ntraps, nplots, and the output value of interest
}
\description{
Rodent data summarized at the relevant level (plot, treatment, site)
}
