% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{observations_are_new}
\alias{observations_are_new}
\title{Find new observations}
\usage{
observations_are_new(base_folder = "~")
}
\arguments{
\item{base_folder}{Folder into which data will be downloaded}
}
\value{
bool True if new observations are available
}
\description{
Check if there are new rodent observations. This only checks the
  Portal_rodent.csv file. (If other data (non-rodent) are updated this function
  will not show that there is new data available.)
}
