% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenocam.R
\name{phenocam}
\alias{phenocam}
\title{Phenocam data products by day, calendar month, or lunar month}
\usage{
phenocam(level = "daily", path = get_default_data_path())
}
\arguments{
\item{level}{specify 'monthly', 'daily', or 'newmoon'}

\item{path}{either the file path that contains the PortalData folder or
"repo", which then pulls data from the PortalData GitHub repository}
}
\description{
Summarize phenocam data products to either daily, monthly, or lunar monthly level.
}
