% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{prep_rodent_output}
\alias{prep_rodent_output}
\title{Rodent data prepared for output}
\usage{
prep_rodent_output(level_data, data_tables, time, effort, na.drop, zero.drop,
  shape, level, output)
}
\arguments{
\item{level_data}{rodent data summarized at the level of interest}

\item{data_tables}{read-in Portal data, passed through for moon data}

\item{time}{specify the format of the time index in the output, either
"period" (sequential Portal surveys), "newmoon" (lunar cycle numbering),
or "date" (calendar date)}

\item{effort}{logical as to whether or not the effort columns should be
included in the output}

\item{na.drop}{logical, drop NA values (representing insufficient sampling)}

\item{zero.drop}{logica, drop 0s (representing sufficient sampling, but no
detections)}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{level}{specify level of interest ("plot", "treatment", "site")}

\item{output}{specify whether to return "abundance", or "biomass", 
or "energy"}
}
\value{
fully crossed flat table of observations with effort (number of
  traps and number of plots). The crossing depends on the level:
  "plot" is period x treatment x plot x species, "treatment" is
  period x treatment x species, and "site" is period x species. Any 
  observations not sufficiently (as defined by min_plots, and 
  hierarchically by min_traps) sampled are returned with NA 
  for ntraps, nplots, and the output value of interest
}
\description{
Rodent data prepared for output
}
