% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{print.portal_data_list}
\alias{print.portal_data_list}
\alias{print}
\title{Prints a portal_data_list object}
\usage{
\method{print}{portal_data_list}(x, ...)
}
\arguments{
\item{x}{A portal_data_list object.}

\item{...}{arguments passed to \code{\link{print}}}
}
\description{
Prints a portal_data_list object
}
