% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_annuals}
\alias{process_annuals}
\title{Restricts species to annuals or non-woody as specified}
\usage{
process_annuals(quadrat_sp_data, type)
}
\arguments{
\item{quadrat_sp_data}{Data table with raw quadrat plant data
merged with species attributes from
species_table.}

\item{type}{String. Either "Annuals" or "Non-woody" results in filtering}
}
\value{
data.table with species processed according to argument 'type'.
}
\description{
If type=Annuals, removes all non-annual species.
             If type=Non-woody, removes shrub species
}
