% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_incomplete_censuses}
\alias{process_incomplete_censuses}
\title{Process incomplete censuses}
\usage{
process_incomplete_censuses(rodent_species_merge, fill_incomplete)
}
\arguments{
\item{rodent_species_merge}{Data table. Merge of raw rodent records and
species information.}

\item{fill_incomplete}{Logical. Denotes if users wants to keep incomplete censuses
or fill with corrected estimates.}
}
\value{
Data.table of merged rodent records and species info with incomplete
        censuses processed according to argument fill_incomplete.
}
\description{
In some months, not all plots are trapped. Using this data can result in
biased monthly data, especially if summarizing for site or treatment.
}
