% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_unknownsp_plants}
\alias{process_unknownsp_plants}
\title{Processes unknown species -- plant data.}
\usage{
process_unknownsp_plants(quadrat_data, unknowns)
}
\arguments{
\item{quadrat_data}{Data.table with raw plant quadrat data.}

\item{unknowns}{String. If unknowns=False, unknown species removed.}
}
\value{
Data.table with species info added and unknown species processed
according to the argument unknowns.
}
\description{
Removes any records for unidentified species if unknowns=FALSE.
If unknowns=TRUE, then their designation in the output file is
given as 'other'.
}
