% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filtering.R
\name{qualityCheck}
\alias{qualityCheck}
\title{Quality check PORTAL freeway data}
\usage{
qualityCheck(fData, tolerance = "high", timeCut)
}
\arguments{
\item{fData}{Your raw freeway dataframe.}

\item{tolerance}{A tolerance level of "high", "medium", or "low". "high" is the least conservative tolerance.}

\item{timeCut}{The time aggregation over which to test quality. Most inuitive strings are accepted, such as "5 min" or "1 hour" or "1 day"}
}
\value{
A list of flags seperated by detector and time period.
}
\description{
Quality check PORTAL freeway data
}

