% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{remove_incomplete_censuses}
\alias{remove_incomplete_censuses}
\title{Remove incomplete censuses}
\usage{
remove_incomplete_censuses(trapping_table, rodent_species_merge, incomplete)
}
\arguments{
\item{trapping_table}{Data table. Data on when each plot was trapped.}

\item{rodent_species_merge}{Data table. Merge of raw rodent records and
species information.}

\item{incomplete}{Boolean. Denotes if users wants to keep incomplete censuses.}
}
\value{
Data.table of merged rodent records and species info with incomplete
        censuses processed according to argument imcomplete.
}
\description{
In some months, not all plots are trapped. Using this data can result in
biased monthly data, especially if summarizing for site or treatment.
}
