% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{remove_suspect_entries}
\alias{remove_suspect_entries}
\title{Remove suspect trapping periods and unknown plots.}
\usage{
remove_suspect_entries(rodent_data)
}
\arguments{
\item{rodent_data}{Data.table of raw rodent data.}
}
\value{
Data.table with suspect data removed.
}
\description{
Removes records with negative period code or 
with missing plot numbers.
}
