% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{rename_species_plants}
\alias{rename_species_plants}
\title{Rename plant species}
\usage{
rename_species_plants(quadrat_data, correct_sp)
}
\arguments{
\item{quadrat_data}{Data.table of raw plant quadrat data.}

\item{correct_sp}{T/F whether or not to use likely corrected plant IDs
[see Methods.md for explanation]}
}
\value{
Data.table with suspected incorrect plant species names replaced
}
\description{
Several species are suspected to have been IDed
incorrectly until 2017, when voucher samples were collected.
    acac greg -> mimo acul
    tali angu -> tali aura
    lcyi torr -> lyci ande
}
