% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodent_species.R
\name{rodent_species}
\alias{rodent_species}
\alias{forecasting_species}
\title{Rodent species abbreviations}
\usage{
rodent_species(
  path = get_default_data_path(),
  type = "code",
  set = "all",
  total = FALSE
)

forecasting_species(
  path = get_default_data_path(),
  total = FALSE,
  type = "abbreviation"
)
}
\arguments{
\item{path}{\code{character} Folder into which data will be downloaded.}

\item{type}{\code{character} value indicating the output type. Current options include \code{'abbreviation'} or \code{'code'} (default, two-letter abbreviation), \code{'g_species'} (abbreviated genus and species), \code{'Latin'} (full scientific names), \code{'common'} (common names), and  \code{'table'} (a \code{data.frame} of all the options).}

\item{set}{\code{character} input of a specified set of species. Options include \code{"all"} (default, all species included) and \code{"forecasting"} (the species used in forecating pipelines).}

\item{total}{\code{logical} value indicating if \code{"total"} should be added or not.}
}
\value{
\code{character} vector of species abbreviations.
}
\description{
Creates a simple \code{character} vector of abbreviations for the Portal \href{http://portal.naturecast.org/profiles.html}{Rodents}.
}
