% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_default_base_folder}
\alias{get_default_base_folder}
\alias{set_default_base_folder}
\title{Manage the default base_folder for downloading Portal Data into}
\usage{
get_default_base_folder(fallback = "~",
  ENV_VAR = "PORTALR_BASE_FOLDER")

set_default_base_folder(base_folder = "~",
  ENV_VAR = "PORTALR_BASE_FOLDER")
}
\arguments{
\item{fallback}{the default value to use if the setting is missing}

\item{ENV_VAR}{the environmental variable to check (by default
`"PORTALR_BASE_FOLDER"``)}

\item{base_folder}{Folder into which data will be downloaded}
}
\value{
None
}
\description{
\code{get_default_base_folder} gets the value of the base_folder
  environmental variable

\code{set_default_base_folder} has 3 steps. First, it checks for
  the presence of a pre-existing setting for the base_folder environmental
  variable. Then it checks if the folder exists and creates it, if needed.
  Then it provides instructions for setting the base_folder environmental
  variable.
}
