% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{shift_current_month}
\alias{shift_current_month}
\title{Shift Current Month}
\usage{
shift_current_month(data)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}
}
\value{

}
\description{
Adjusts month for a diff pattern of: prev_month = 2, focal_month=0, next_month=1
which matches an example trapping pattern of Jan trapped, Feb not trapped, Mar trapped twice,
Apr trapped where the early march trapping is just late for February.
}
