% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{shift_current_next}
\alias{shift_current_next}
\title{Shift Current and Next Month}
\usage{
shift_current_next(data)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}
}
\value{

}
\description{
# Adjusts month for a diff pattern of: prev_month = 2, focal_month=1, next_month=0
which matches an example trapping pattern of January trapped, February missed, March trapped,
April trapped twice. In this scenario, the February census was conducted in early March and March
census was conducted in early April.
}
