% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthly_timeseries.R
\name{shift_next_month}
\alias{shift_next_month}
\title{Shift Next Month}
\usage{
shift_next_month(data)
}
\arguments{
\item{data}{Dataframe with columns date (date of the period (e.g. 2016-01-01)), period
(period code for the census) and value (numeric value to be analyzed)}
}
\value{

}
\description{
Adjusts month for a diff pattern of: prev_month = 1, focal_month=0, next_month=2
which matches an example trapping pattern of January trapped, February trapped twice, March
missed where the late February census was early for March.
}
