% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{true_length}
\alias{true_length}
\title{Length of non-missing values in a vector

How many values in the given vector are not NAs}
\usage{
true_length(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
integer of how many non-NA values in x
}
