% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_function.R
\name{add_seasons}
\alias{add_seasons}
\title{Add Seasons}
\usage{
add_seasons(data, num_of_seasons)
}
\arguments{
\item{data}{data frame containing columns: yr, mo, species}

\item{num_of_seasons}{either 2: winter = Oct-March
      summer = April-Sept
or 4: winter = Dec-Feb
      spring = March-May
      summer = Jun-Aug
      fall = Sep-Nov}
}
\description{
Adds seasons to data frame
}
