% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{check_for_newer_data}
\alias{check_for_newer_data}
\title{Check for latest version of data files}
\usage{
check_for_newer_data(path = get_default_data_path())
}
\arguments{
\item{path}{Folder in which data will be checked}
}
\value{
bool TRUE if there is a newer version of the data online
}
\description{
Check the latest version against the data that exists on
  the GitHub repo
}
