% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{download_observations}
\alias{download_observations}
\title{Download the PortalData repo}
\usage{
download_observations(base_folder = "~", release_only = TRUE)
}
\arguments{
\item{base_folder}{Folder into which data will be downloaded}

\item{release_only}{whether to download the "Release" version of the data. Set to FALSE
to download the most up to date data.}
}
\value{
None
}
\description{
This downloads the latest portal data regardless if they are
  actually updated or not.
  TODO: incorporate data retriever into this when it's pointed at the github repo
}
