% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{download_observations}
\alias{download_observations}
\title{Download the PortalData repo}
\usage{
download_observations(base_folder = "~", version = "latest",
  from_zenodo = TRUE)
}
\arguments{
\item{base_folder}{Folder into which data will be downloaded}

\item{version}{Version of the data to download (default = "latest")}

\item{from_zenodo}{logical; if `TRUE`, get info from Zenodo, otherwise GitHub}
}
\value{
None
}
\description{
This downloads the latest portal data regardless if they are
  actually updated or not.
  TODO: incorporate data retriever into this when it's pointed at the github repo
}
\examples{
\donttest{
  download_observations()
  download_observations("~/old-data", version = "1.50.0")
}

}
