% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather.R
\name{fill_missing_weather}
\alias{fill_missing_weather}
\title{Fill missing weather with regional data}
\usage{
fill_missing_weather(weather, path = "~")
}
\arguments{
\item{weather}{a dataframe of daily weather data}

\item{path}{specify where to locate regional data}
}
\description{
Use two weather stations in San Simon valley to fill in
missing weather data in the daily time series
}
