% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{fill_weight}
\alias{fill_weight}
\title{Fill Weight}
\usage{
fill_weight(rodent_data, tofill)
}
\arguments{
\item{rodent_data}{raw rodent data}

\item{tofill}{logical whether to fill in missing values or not}
}
\description{
fill in missing weight values with either a recently recorded
  weight for that individual or species average
}
