% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rodent_data.R
\name{find_incomplete_censuses}
\alias{find_incomplete_censuses}
\title{Period code for incomplete censuses}
\usage{
find_incomplete_censuses(trapping_table, min_plots, min_traps)
}
\arguments{
\item{trapping_table}{Data_table of when plots were censused.}

\item{min_plots}{minimum number of plots within a period for an
observation to be included}

\item{min_traps}{minimum number of traps for a plot to be included}
}
\value{
Data.table of period codes when not all plots were trapped.
}
\description{
Determines incomplete censuses by finding dates when some plots
  were trapped, but others were not.
}
