% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{find_incomplete_censuses}
\alias{find_incomplete_censuses}
\title{Period code for incomplete censuses}
\usage{
find_incomplete_censuses(trapping_table)
}
\arguments{
\item{trapping_table}{Data table. Data on when each plot was trapped.}
}
\value{
Data.table of period codes when not all plots were trapped.
}
\description{
Determines incomplete censuses by finding dates when some plots were trapped, but others were not.
}
