% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{FullPath}
\alias{FullPath}
\title{Return normalized path for all operating systems}
\usage{
FullPath(ReferencePath, BasePath = getwd())
}
\arguments{
\item{ReferencePath}{a path to join with current working directory}

\item{BasePath}{Current working directory else path given}
}
\description{
Return normalized path for all operating systems
}
\examples{
FullPath('PortalData/Rodents/Portal_rodent.csv')
FullPath('PortalData/Rodents/Portal_rodent.csv', '~')
}
