% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantData.R
\name{get_plant_data}
\alias{get_plant_data}
\alias{plant_abundance}
\title{Generate summaries of Portal plant data}
\usage{
get_plant_data(path = "~", level = "Site", type = "All", plots = "all",
  unknowns = FALSE, correct_sp = TRUE, shape = "flat",
  output = "abundance")

plant_abundance(..., shape = "flat")
}
\arguments{
\item{path}{path to location of downloaded Portal data; or "repo" to
retrieve data from github repo}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species;
If type=Annuals, removes all non-annual species.
If type=Summer Annuals, returns all annual species that can be found in the summer
If type=Winter Annuals, returns all annual species that can be found in the winter
If type=Non-woody, removes shrub and subshrub species
If type=Perennials, returns all perennial species (includes shrubs and subshrubs)
If type=Shrubs, returns only shrubs and subshrubs}

\item{plots}{specify subset of plots; can be a vector of plots, or specific
sets: "all" plots or "Longterm" plots (plots that have had the same
treatment for the entire time series)}

\item{unknowns}{either removes all individuals not identified to species
(unknowns = FALSE) or sums them in an additional column (unknowns = TRUE)}

\item{correct_sp}{correct species names suspected to be incorrect in early data (T/F)}

\item{shape}{return data as a "flat" list or "crosstab"}

\item{output}{specify whether to return "abundance", or "cover" [cover data
starts in summer 2015]}

\item{...}{arguments passed to \code{\link{get_plant_data}}}
}
\value{
a data.frame in either "long" or "wide" format, depending on the
  value of `shape`
}
\description{
This function is a generic interface into creating
  summaries of the Portal plant species data. It contains a number of
  arguments to specify both the kind of data to summarize, at what level of
  aggregation, various choices for dealing with data quality, and output
  format.

\code{plant_abundance} generates a table of plant abundance
}
\examples{
plant_abundance("repo")

}
