% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{get_rodent_data}
\alias{get_rodent_data}
\alias{abundance}
\alias{biomass}
\alias{abundance}
\alias{biomass}
\title{Retrieve Rodent Data}
\usage{
get_rodent_data(path = "~", level = "Site", type = "Rodents",
  length = "all", unknowns = FALSE, incomplete = FALSE,
  shape = "crosstab", time = "period", output = "abundance")

abundance(...)

biomass(...)
}
\arguments{
\item{path}{path to location of downloaded Portal data; or 'repo' to retrieve data from github repo}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species; either all "Rodents" or only "Granivores"}

\item{length}{specify subset of plots; use "All" plots or only "Longterm" plots (plots that have had same treatment for entire time series)}

\item{unknowns}{either removes all individuals not identified to species (unknowns=F) or sums them in an additional column (unknowns=T)}

\item{incomplete}{either removes all data from incomplete trapping sessions (incomplete = F) or includes them (incomplete = T)
[note that if level="plot" and incomplete=T, NAs will be included in periods where trapping was incomplete]}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{time}{return data using the complete "newmoon" numbers of the original "period" numbers}

\item{output}{specify whether to return "abundance", or "biomass"}
}
\description{
A function to summarize monthly Portal rodent species data.

\code{abundance} essentially passes along all arguments to \code{get_rodent_data}, but fixes the output type as "abundance"

\code{biomass} essentially passes along all arguments to \code{get_rodent_data}, but fixes the output type as "biomass"
}
\examples{
abundance("repo")

biomass("repo")

}
