% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RodentData.R
\name{get_rodent_data}
\alias{get_rodent_data}
\alias{abundance}
\alias{biomass}
\alias{energy}
\alias{abundance}
\alias{biomass}
\alias{energy}
\title{Generate summaries of Portal rodent data}
\usage{
get_rodent_data(path = "~", clean = TRUE, level = "Site",
  type = "Rodents", length = "all", plots = length, unknowns = FALSE,
  shape = "crosstab", time = "period", output = "abundance",
  fillweight = (output != "abundance"), na_drop = switch(tolower(level),
  plot = FALSE, treatment = TRUE, site = TRUE),
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site =
  TRUE), min_traps = 1, min_plots = 24, effort = FALSE,
  download_if_missing = TRUE)

abundance(...)

biomass(...)

energy(...)
}
\arguments{
\item{path}{path to location of downloaded Portal data; or "repo" to
retrieve data from github repo}

\item{clean}{logical, load only QA/QC rodent data (TRUE) or all data (FALSE)}

\item{level}{summarize by "Plot", "Treatment", or "Site"}

\item{type}{specify subset of species; either all "Rodents" or only
"Granivores"}

\item{length}{specify subset of plots; use "All" plots or only "Longterm"
plots (to be deprecated)}

\item{plots}{specify subset of plots; can be a vector of plots, or specific
sets: "all" plots or "Longterm" plots (plots that have had the same
treatment for the entire time series)}

\item{unknowns}{either removes all individuals not identified to species
(unknowns = FALSE) or sums them in an additional column (unknowns = TRUE)}

\item{shape}{return data as a "crosstab" or "flat" list}

\item{time}{specify the format of the time index in the output, either
"period" (sequential Portal surveys), "newmoon" (lunar cycle numbering),
"date" (calendar date)}

\item{output}{specify whether to return "abundance", or "biomass", or
"energy"}

\item{fillweight}{specify whether to fill in unknown weights with other
records from that individual or species, where possible}

\item{na_drop}{logical, drop NA values (representing insufficient sampling)}

\item{zero_drop}{logical, drop 0s (representing sufficient sampling, but no
detections)}

\item{min_traps}{minimum number of traps for a plot to be included}

\item{min_plots}{minimum number of plots within a period for an
observation to be included}

\item{effort}{logical as to whether or not the effort columns should be
included in the output}

\item{download_if_missing}{if the specified file path doesn't have the
PortalData folder, then download it}

\item{...}{arguments passed to \code{\link{get_rodent_data}}}
}
\value{
a data.frame in either "long" or "wide" format, depending on the
  value of `shape`
}
\description{
This function is a generic interface into creating summaries
  of the Portal rodent species data. It contains a number of arguments
  to specify the kind of data to summarize (at what level of aggregation)
  and various choices for dealing with data quality, and output format.

\code{abundance} generates a table of rodent abundance

* \code{biomass()} generates a table of rodent biomass

* \code{energy()} generates a table of rodent energy
  (computed as biomass ^ 0.75)
}
\examples{
abundance("repo")

biomass("repo")

energy("repo")

}
