% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_trapping_to_rodents}
\alias{join_trapping_to_rodents}
\title{Join rodent and trapping tables}
\usage{
join_trapping_to_rodents(rodent_data, trapping_table, incomplete)
}
\arguments{
\item{rodent_data}{Data.table with raw rodent data.}

\item{trapping_table}{Data_table of when plots were censused.}

\item{incomplete}{Boolean. Denotes if users wants to keep incomplete censuses.}
}
\value{
Data.table of raw rodent data with trapping info added.
}
\description{
Joins rodent data with list of trapping dates, by period and plot
}
