% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{join_trapping_to_rodents}
\alias{join_trapping_to_rodents}
\title{Join rodent and trapping tables}
\usage{
join_trapping_to_rodents(rodent_data, trapping_table, full_trapping, min_plots,
  min_traps)
}
\arguments{
\item{rodent_data}{Data.table with raw rodent data.}

\item{trapping_table}{Data_table of when plots were censused.}

\item{full_trapping}{Unfiltered data_table of when plots were censused.}

\item{min_plots}{minimum number of plots within a period for an
observation to be included}

\item{min_traps}{minimum number of plots within a period for an
observation to be included}
}
\value{
Data.table of raw rodent data with trapping info added.
}
\description{
Joins rodent data with list of trapping dates, by period and plot
}
