% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_data}
\alias{load_data}
\title{Loads the Portal rodent data files}
\usage{
load_data(path = "~")
}
\arguments{
\item{path}{either the file path that contains the PortalData folder or
"repo", which then pulls data from the PortalData GitHub repository}
}
\value{
List of 5 dataframes:
  \itemize{
    \item rodent_data. raw data on rodent captures
    \item species_table. species code, names, types
    \item trapping_table. when each plot was trapped
    \item newmoons_table. pairs census periods with newmoons
    \item plots_table. rodent treatment assignments for each plot.
  }
}
\description{
Loads main Portal rodent data files from either a user-defined
  path or the online Github repository. If the user-defined path is un-
  available, the default option is to download to that location.
}
\examples{
portal_data <- load_data("repo")

}
