% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{make_crosstab}
\alias{make_crosstab}
\title{Make Crosstab}
\usage{
make_crosstab(summary_data)
}
\arguments{
\item{summary_data}{summarized rodent data - must include 'abundance' column}
}
\description{
convert summarized rodent data to crosstab form
}
