% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{observations_are_new}
\alias{observations_are_new}
\title{Check if there are new rodent observations. This only checks the
Portal_rodent.csv file. If other things are updated this function
will not show that there is new data available.}
\usage{
observations_are_new(base_folder = "~")
}
\value{
bool True if new observations are available
}
\description{
Check if there are new rodent observations. This only checks the
Portal_rodent.csv file. If other things are updated this function
will not show that there is new data available.
}
