% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather.R
\name{weather}
\alias{weather}
\title{Weather by day, calendar month, or lunar month}
\usage{
weather(level = "daily", fill = FALSE,
  path = get_default_data_path())
}
\arguments{
\item{level}{specify 'monthly', 'daily', or 'newmoon'}

\item{fill}{specify if missing data should be filled, passed to \code{fill_missing_weather}}

\item{path}{either the file path that contains the PortalData folder or
"repo", which then pulls data from the PortalData GitHub repository}
}
\description{
Summarize hourly weather data to either daily, monthly, or lunar monthly level.
}
