% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weather.R
\name{weather}
\alias{weather}
\title{Weather by day or month}
\usage{
<<<<<<< HEAD
weather(level = "daily", fill = FALSE, path = "~")
}
\arguments{
\item{level}{specify 'monthly' or 'daily'}
=======
weather(level = "Daily", fill = FALSE, path = "~")
}
\arguments{
\item{level}{specify 'Monthly' or 'Daily'}

\item{fill}{specify if missing data should be filled, passed to \code{fill_missing_weather}}

\item{path}{specify where to locate Portal data}
>>>>>>> upstream/fill_weather

\item{fill}{specify if missing data should be filled, passed to \code{fill_missing_weather}}

\item{path}{specify where to locate Portal data}
}
\description{
Summarize hourly weather data to either daily or monthly level.
}
